+++ apps/client/src/app/components/portfolio-summary/portfolio-summary.component.ts
@@ -1,5 +1,6 @@
 import { getDateFnsLocale, getLocale } from '@ghostfolio/common/helper';
 import { PortfolioSummary } from '@ghostfolio/common/interfaces';
+import { translate } from '@ghostfolio/ui/i18n';
 
 import {
   ChangeDetectionStrategy,
@@ -28,6 +29,9 @@ export class PortfolioSummaryComponent implements OnChanges, OnInit {
 
   @Output() emergencyFundChanged = new EventEmitter<number>();
 
+  public buyAndSellActivitiesTooltip = translate(
+    'BUY_AND_SELL_ACTIVITIES_TOOLTIP'
+  );
   public timeInMarket: string;
 
   public constructor() {}