+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Moved the holdings table to the holdings tab of the home page
 - Improved the performance labels (with and without currency effects) in the position detail dialog
 
+### Fixed
+
+- Fixed an issue with the benchmarks in the markets overview
+- Fixed an issue with the _Fear & Greed Index_ (market mood) in the markets overview
+
 ## 2.78.0 - 2024-05-02
 
 ### Added